requirements.txt
Jinja2==3.1.2
PyGithub==1.55
pygrok==1.0.0
pandas==1.5.0
mpld3==0.5.8
plotly==5.10.0
nbformat==5.6.0